requirements.txt
absl_py==1.3.0
dlib==19.24.0
numpy==1.23.5
opencv_python==4.6.0.66
